refactor: bookRecordDataProps 분리
