+++ src/types/record.d.ts
@@ -1,0 +1,15 @@
+export interface bookRecordDataProps {
+  record: {
+    title: string;
+    thumbnail: string;
+    bookIsbn: string;
+    text: string;
+    recordImg: string;
+    recordImgUrl: string;
+    tags?: {
+      id: number;
+      data: string | number;
+    }[];
+    readAt: string;
+  };
+}
